refactor: 앨범/스포티파이 관련 api들 album.ts, spotify.ts로 이동
