+++ app/(routes)/[genre]/page.tsx
@@ -1,7 +1,7 @@
 import Error from "../../components/@common/Error";
 import GenreContents from "../../components/@common/GenreContents";
 import { MusicLayout } from "../../components/@common/MusicLayout";
-import { fetchGenreData } from "../../modules/api";
+import { fetchGenreData } from "../../modules/api/album";
 import { FIRST_PAGE } from "../../modules/constants";
 import { PageProps } from "../../modules/types";
