+++ app/(routes)/admin/search/[keyword]/[page]/page.tsx
@@ -1,7 +1,7 @@
 import SearchContents from "@/app/components/search/SearchContents";
 import { MusicLayout } from "@/app/components/@common/MusicLayout";
 import { PageProps } from "@/app/modules/types";
-import { fetchSearchData } from "@/app/modules/api";
+import { fetchSearchData } from "@/app/modules/api/album";
 
 export default async function Page({ params }: PageProps) {
   const currentKeyword: string = params.keyword;
