+++ app/(routes)/artist/[id]/[page]/page.tsx
@@ -1,7 +1,7 @@
 import Error from "@/app/components/@common/Error";
 import ArtistContents from "../../../../components/artist/ArtistContents";
 import { MusicLayout } from "@/app/components/@common/MusicLayout";
-import { fetchArtistData } from "@/app/modules/api";
+import { fetchArtistData } from "@/app/modules/api/album";
 import { SITE_TITLE } from "@/app/modules/constants";
 import { PageProps } from "@/app/modules/types";
 import React from "react";
