+++ app/(routes)/page.tsx
@@ -1,7 +1,7 @@
 import Error from "../components/@common/Error";
 import { MusicLayout } from "../components/@common/MusicLayout";
 import { LandingPage } from "../components/landingPage/LandingPage";
-import { fetchAlbumDataSSR } from "../modules/api";
+import { fetchAlbumDataSSR } from "../modules/api/album";
 import { PER_PAGE_COUNT } from "../modules/constants";
 
 export default async function Page() {
