+++ app/(routes)/post/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { MusicLayout } from "../../../components/@common/MusicLayout";
 import { Metadata } from "next/dist/lib/metadata/types/metadata-interface";
 import { PageProps } from "../../../modules/types";
-import { fetchPostData } from "../../../modules/api";
+import { fetchPostData } from "../../../modules/api/album";
 import { SITE_TITLE } from "@/app/modules/constants";
 import Error from "@/app/components/@common/Error";
 import { PostContents } from "@/app/components/post/PostContents";
