+++ app/(routes)/search/[keyword]/[page]/page.tsx
@@ -1,7 +1,7 @@
 import SearchContents from "@/app/components/search/SearchContents";
 import { MusicLayout } from "@/app/components/@common/MusicLayout";
 import { PageProps } from "@/app/modules/types";
-import { fetchSearchData } from "@/app/modules/api";
+import { fetchSearchData } from "@/app/modules/api/album";
 import Error from "@/app/components/@common/Error";
 
 export default async function Page({ params }: PageProps): Promise<React.ReactElement> {
