+++ app/components/@common/header/MobileMenu.tsx
@@ -3,7 +3,7 @@ import styles from "./MobileMenu.module.css";
 import Link from "next/link";
 import { toCalendarPage, toGenrePage, toPostPage } from "@/app/modules/paths";
 import { usePathname, useRouter } from "next/navigation";
-import { fetchRandomAlbumId } from "@/app/modules/api";
+import { fetchRandomAlbumId } from "@/app/modules/api/album";
 import { isAdminPage } from "@/app/modules/utils";
 import { useEffect } from "react";
 import { userImageAtom, userNameAtom } from "@/app/modules/atoms";
