+++ app/modules/api/spotify.ts
@@ -1,0 +1,42 @@
+export const fetchSpotify = async (albumId: string) => {
+  if (!albumId) {
+    alert("모든 항목을 채워주세요.");
+    return;
+  }
+
+  try {
+    const queryString = `?albumId=${albumId}`;
+    const url = `/api/spotify/fetch${queryString}`;
+
+    const response = await fetch(url, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+
+    const fetchedData = await response.json();
+    return fetchedData;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+export const searchSpotify = async (searchKeyword: string) => {
+  try {
+    const queryString = `?searchKeyword=${searchKeyword}`;
+    const url = `/api/spotify/search${queryString}`;
+
+    const response = await fetch(url, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    });
+
+    const searchData = await response.json();
+    return searchData;
+  } catch (error) {
+    console.error(error);
+  }
+};
